updated Readme and example APIs
