+++ examples/README.md
@@ -1,14 +1,5 @@
 # Fluvio Client Node Examples
 
-## Setup To run any of these examples make sure you've got a valid node setup:
-
-If you clone this repository navigate to this directory in the repository and run:
-```node
-npm init -y
-npm install -D typescript ts-node @types/node
-npm install -S @fluvio/client
-```
-
 ## Prerequisites
 
 Before running the examples, you'll need to have completed the following:
@@ -24,6 +15,19 @@ Before running the examples, you'll need to have completed the following:
 [Fluvio CLI]: https://fluvio.io/download
 [local cluster]: https://fluvio.io/docs
 
+## Create a Node Project
+
+In the `examples` directory, create a node project and install the following libraries:
+
+```node
+npm init -y
+npm install -D typescript ts-node @types/node
+npm install -S @fluvio/client
+```
+
+Note: Fluvio is installed from `npi` rather than called locally.
+
+
 ## Examples
 
 All examples run with:
@@ -36,6 +40,10 @@ All examples run with:
 * [Find Topic](#find-topic)
 * [List Topics](#list-topics)
 * [Delete Topic](#delete-topic)
+* [Produce](#produce)
+* [Produce Batch](#produce-batch)
+* [Consume](#consume)
+* [Consume Iterator](#consume-iterator)
 
 
 ### Create Topic
@@ -133,14 +141,96 @@ connecting client to fluvio
 topic 'node-examples' deleted
 ```
 
+### Produce
 
-Find the topic `node-examples`:
+Produce 10 records to `node-example` topic.
 
-To run the a given example do `npx ts-node ./<name-of-example>.ts` such as:
-* `npx ts-node ./consume.ts`
-* `npx ts-node ./iterator.ts`
-* `npx ts-node ./produce.ts`
-* `npx ts-node ./produceBatch.ts`
+```
+% npx ts-node ./produce.ts
+connecting client to fluvio
+{"key":1,"message":"Value 1"}
+{"key":2,"message":"Value 2"}
+{"key":3,"message":"Value 3"}
+{"key":4,"message":"Value 4"}
+{"key":5,"message":"Value 5"}
+{"key":6,"message":"Value 6"}
+{"key":7,"message":"Value 7"}
+{"key":8,"message":"Value 8"}
+{"key":9,"message":"Value 9"}
+{"key":10,"message":"Value 10"}
+```
+
+The producer example shows 2 API variants, one with `key` and the other without.
+
+### Produce Batch
+
+Produce 10 records to `node-example` topic using `batching`. Batching small packets can significantly improve performance.
+
+```
+% npx ts-node ./produceBatch.ts
+connecting client to fluvio
+[
+  [
+    Uint8Array(5) [ 75, 69, 89, 32, 49 ],
+    '{"key":1,"message":"Message 1"}'
+  ],
+  [
+    Uint8Array(5) [ 75, 69, 89, 32, 50 ],
+    '{"key":2,"message":"Message 2"}'
+  ],
+  ...
+]
+```
+
+Checkout the `sendAll` API for additional details.
+
+### Consume
+
+Consume all records stored in `node-example`:
+
+```
+% npx ts-node ./consume.ts
+connecting client to fluvio
+read from beginning
+Key=KEY 1, Value={"key":1,"message":"Value 1"}
+Key=KEY 2, Value={"key":2,"message":"Value 2"}
+Key=KEY 3, Value={"key":3,"message":"Value 3"}
+Key=KEY 4, Value={"key":4,"message":"Value 4"}
+Key=KEY 5, Value={"key":5,"message":"Value 5"}
+Key=KEY 6, Value={"key":6,"message":"Value 6"}
+Key=KEY 7, Value={"key":7,"message":"Value 7"}
+Key=KEY 8, Value={"key":8,"message":"Value 8"}
+Key=KEY 9, Value={"key":9,"message":"Value 9"}
+Key=KEY 10, Value={"key":10,"message":"Value 10"}
+^C
+```
+
+The consume reads all records from the beginning of list, then waits for new records. The offset controls the location of the pointer in the list. Checkout the API for additional settings.
 
 
-Produce/Consume requires a topic called `node-examples`. Use `createTopic` example to created it, and `deleteTopic` to remove it.
+### Consume Iterator
+
+As oposed to the previous `Consume` API, the conume iterator reads to completion. The API exists when it reches the end of the list.
+
+Consumer reads all messages from the beginning of `node-example`:
+
+```
+% npx ts-node ./consumeIterator.ts
+connecting client to fluvio
+read from node-examples
+Key=KEY 1, value={"key":1,"message":"Value 1"}
+Key=KEY 2, value={"key":2,"message":"Value 2"}
+Key=KEY 3, value={"key":3,"message":"Value 3"}
+Key=KEY 4, value={"key":4,"message":"Value 4"}
+Key=KEY 5, value={"key":5,"message":"Value 5"}
+Key=KEY 6, value={"key":6,"message":"Value 6"}
+Key=KEY 7, value={"key":7,"message":"Value 7"}
+Key=KEY 8, value={"key":8,"message":"Value 8"}
+Key=KEY 9, value={"key":9,"message":"Value 9"}
+Key=KEY 10, value={"key":10,"message":"Value 10"}
+```
+
+Note, in this example the API retrieves both the key and the value.
+
+## Other references
+* [Typedoc API page](https://infinyon.github.io/fluvio-client-node/)
