+++ examples/consumeIterator.ts
@@ -4,8 +4,8 @@ import Fluvio, {Offset, Record} from "@fluvio/client";
 const TOPIC_NAME = 'node-examples'
 const PARTITION = 0
 
-async function consume() {
-    try {
+async function consumeIterator() {
+    try {    
         const fluvio = new Fluvio()
 
         console.log('connecting client to fluvio')
@@ -16,16 +16,22 @@ async function consume() {
         // Create partition consumer
         const consumer = await fluvio.partitionConsumer(TOPIC_NAME, PARTITION)
 
-        console.log('read from the beginning')
+        console.log(`read from ${TOPIC_NAME}`)
 
-        await consumer.stream(Offset.FromBeginning(), async (record: Record) => {
-            // handle record;
-            console.log(`Key=${record.keyString()}, Value=${record.valueString()}`)
-        })
+        let count = 1
+        let stream = await consumer.createStream(Offset.FromBeginning())
+
+        for await (const record of stream) {
+            const key = record.keyString()
+            const value = record.valueString()
+            console.log(`Key=${key}, value=${value}`)
+            if (count >= 10) break
+            count++
+        }
 
     } catch (ex) {
         console.log('error', ex)
-    }
+    }        
 }
 
-consume()
+consumeIterator()
