+++ examples/produceBatch.ts
@@ -1,5 +1,5 @@
 /* tslint:disable:no-console */
-import Fluvio from "@fluvio/client";
+import Fluvio, {KeyValue} from "@fluvio/client";
 
 const TOPIC_NAME = 'node-examples'
 
@@ -14,21 +14,24 @@ async function produceBatch() {
 
         const producer = await fluvio.topicProducer(TOPIC_NAME)
         const records: KeyValue[] = []
-        for (let i: number = 1; i < 10; i++) {
+        for (let i: number = 1; i <= 10; i++) {
+
             // Create a JSON message as our value
             const message = JSON.stringify({
                 key: i,
-                message: `This is message ${i}`,
+                message: `Message ${i}`,
             })
 
             const encoder = new TextEncoder()
             const key: ArrayBuffer = encoder.encode(`KEY ${i}`)
             records.push([key, message])
         }
-
+        
         console.log(records);
 
+        // send all records in a batch
         await producer.sendAll(records)
+
     } catch (ex) {
         console.log('error', ex)
     }
